fix: use biginteger over integers for ids

+++ mp2i/database/models/guild.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from . import Base
@@ -12,7 +12,7 @@ class Guild(Base):
     __tablename__ = "guilds"
 
     guild_id: Mapped[int] = mapped_column(
-        Integer(),
+        BigInteger(),
         primary_key=True,
         nullable=False,
         info=dict(
